move BeanValidationConfigConsistencyChecker to core and enable it by default (via ConfigConsistencyCheckerImpl)
