+++ matsim/src/main/java/org/matsim/core/config/consistency/ConfigConsistencyCheckerImpl.java
@@ -52,6 +52,7 @@ public final class ConfigConsistencyCheckerImpl implements ConfigConsistencyChec
 		checkPlanCalcScore(config);
 		checkTransit(config);
 		checkConsistencyBetweenRouterAndTravelTimeCalculator( config );
+		new BeanValidationConfigConsistencyChecker().checkConsistency(config);
 	}
 
 	static boolean checkConsistencyBetweenRouterAndTravelTimeCalculator( final Config config ) {
